refactor(voice): extract TTSServiceSupervisor to tts_service.py (issue #94 phase 3)
把「服务进程监督」职责从 GPTSoVITSTTSProvider 抽到新模块 tts_service.py：
- TTSServiceSupervisor / GenieServiceSupervisor：探测/启动/接管/Broken pipe 重启/
  角色权重/模型加载，状态仍挂在 self 上（_server_process/_service_checked/
  _weights_ready/_service_state），鸭子桩测试只需换类名、桩字段不变
- 本地子进程经 ProcessResource 托管，detach_local_service 交出所有权不杀进程，
  close() 改走协调器自持 RM 的 stop_all 统一关闭
- 协调器持有 supervisor 并委托 service_ready/ensure_ready/detach_local_service，
  合成路径的 _ensure_service_available/_ensure_character_weights/_restart_* 改委托；
  settings 改为只读 property 委托 supervisor，使 Genie 备用端口切换传播到合成
- 迁移配套的进程/探测/URL/Genie 模型 helper；tts.py re-export 供既有导入
- 消除 GenieTTSProvider 对服务方法的继承覆写（仅保留 _create_supervisor +
  _request_audio），协调器级继承待提交 6 收尾
- 重写 test_tts.py 服务段（类名指向 supervisor、monkeypatch 目标改 tts_service）

回归：tests/unit + tests/ui 全绿（988 passed，排除 2 个已知环境性失败）。
